% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{usses_ibutton}
\alias{usses_ibutton}
\title{Daily ibutton temperature data from rainfall shelters}
\format{A data frame with 6591 rows and 5 variables:
\describe{
  \item{date}{date in \%Y-\%m-\%d format}
  \item{plot}{experimental plot match with usses_quads}
  \item{n}{ number of samples recorded on a given date}
  \item{stat}{daily temperature measure, TMAX, TMIN, TMEAN}
  \item{ibutton}{recorded value in degrees C}
}}
\usage{
usses_ibutton
}
\description{
A dataset containing the date, the daily max, min and mean
mean temperatures recorded by iButtons placed 50 cm above ground
level at the north end of precipitation and irrigation plots.
iButtons recorded data every two hours and this was aggregated by
day to daily tmax, tmin and tmean ((tmax+tmin)/2). iButtons were
housed in short sections of pvc pipe to reduce effects of
direct sun exposure.
}
\keyword{datasets}
