% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{usses_quads}
\alias{usses_quads}
\title{Information on USSES monitoring quadrats}
\format{A data frame with 70 rows and 7 variables:
\describe{
  \item{QuadName}{character giving full quadrat name}
  \item{quad}{short code for each quadrat starting with "q"}
  \item{Grazing}{indicator for quadrats that recieve grazing from sheep}
  \item{Group}{sptial group for each quadrat}
  \item{paddock}{large pasture where quadrats are located, may be grazed or ungrazed}
  \item{Treatment}{Experimental treatment. Rainfall and competition were manipulated in some quadrats}
  \item{PrecipGroup}{Grouping for precipitation experiment.  Experimental plots were established in groups of three.}
}}
\usage{
usses_quads
}
\description{
Historical and experimental 1 x 1 m quadrats
}
\keyword{datasets}
