% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{usses_soilwat}
\alias{usses_soilwat}
\title{Daily soil moisture data modeled by SOILWAT}
\format{A data frame with 33141 rows and 8 variables:
\describe{
  \item{Year}{year}
  \item{DOY}{Day of year, 1-365}
  \item{Lyr_1}{Estimated volumetric soil moisture in layer}
  \item{Lyr_2}{Estimated volumetric soil moisture in layer}
  \item{Lyr_3}{Estimated volumetric soil moisture in layer}
  \item{Lyr_4}{Estimated volumetric soil moisture in layer}
  \item{Lyr_5}{Estimated volumetric soil moisture in layer}
  \item{Lyr_6}{Estimated volumetric soil moisture in layer}
}}
\usage{
usses_soilwat
}
\description{
Daily soil moisture modeled from daily weather input using
the SOILWAT model maintained by the Burke-Lauenrothe lab.
rSOILWAT2 package is available at https://github.com/DrylandEcology/rSOILWAT2
}
\keyword{datasets}
