% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{usses_spot_sm}
\alias{usses_spot_sm}
\title{Measures of soil moisture from around experimental monitoring plots}
\format{A data frame with 624 rows and 10 variables:
\describe{
  \item{plot}{unique plot id, matches QuadName in quad_info }
  \item{VWC}{Volumetric water content}
  \item{date}{Date in \%Y-\%m-\%d format}
  \item{rep}{replicate measurement of soil moisture, indicates position around the plot}
  \item{quad}{short code for each quadrat starting with "q"}
  \item{Grazing}{indicator for quadrats that recieve grazing from sheep}
  \item{Group}{sptial group for each quadrat}
  \item{paddock}{large pasture where quadrats are located, may be grazed or ungrazed}
  \item{Treatment}{Experimental treatment. Rainfall and competition were manipulated in some quadrats}
  \item{PrecipGroup}{Grouping for precipitation experiment.  Experimental plots were established in groups of three.}
}}
\usage{
usses_spot_sm
}
\description{
Soil moisture was measured using decagon probes around each of the experimental
plots in the spring.
}
\keyword{datasets}
